added tilemap
added a tilemap and configured tiles, added TilemapSetup scene to showcase how it works with an example.

the automapping isn't 100% but it provides a good base to start from that you can easily edit. to edit further, click the "Disable autotile" checkbox in the tilemap window.

let me know if you guys want a demo of how I used this!